photo_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-agent"
version = "1.0.7"
description = "AI-powered image processing pipeline for e-commerce photography"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Photo Agent Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
keywords = ["image-processing", "ai", "computer-vision", "ecommerce", "photography"]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "Pillow>=9.0.0",
    "rembg>=2.0.0",
    "requests>=2.25.0",
    "transformers>=4.20.0",
    "torch>=1.9.0",
    "gdown>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.scripts]
photo-agent = "photo_agent.src.cli:main"

[project.urls]
Homepage = "https://github.com/example/photo-agent"
Documentation = "https://github.com/example/photo-agent#readme"
Repository = "https://github.com/example/photo-agent.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["photo_agent*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=photo_agent"
